requirements.txt
absl-py==0.7.1
aenum==2.1.2
astor==0.8.0
attrs==19.1.0
Click==7.0
click-plugins==1.1.1
cligj==0.5.0
cycler==0.10.0
Django==2.2.2
Fiona==1.8.6
gast==0.2.2
GDAL==2.4.1
geopandas==0.5.0
grpcio==1.21.1
h5py==2.9.0
html5lib==1.0.1
joblib==0.13.2
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
matplotlib==3.0.3
mock==3.0.5
munch==2.3.2
numpy==1.16.4
pandas==0.24.2
Pillow==6.0.0
protobuf==3.8.0
psycopg2==2.8.2
pyparsing==2.4.0
PyPDF2==1.26.0
pyproj==2.2.0
python-dateutil==2.8.0
pytz==2019.1
PyYAML==5.1
reportlab==3.5.23
scikit-learn==0.21.2
scipy==1.3.0
seaborn==0.9.0
Shapely==1.6.4.post1
six==1.12.0
sklearn==0.0
sqlparse==0.3.0
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
webencodings==0.5.1
Werkzeug==0.15.4
xhtml2pdf==0.2.3
